Refactor projects to appear on the index page.
Place the project list display css into it's own separate folder
and use the project list csv to make a project list on both
the projects page and the index for minimal code repetition.

The project list on the index page is composed solely of projects
marked in the csv list by `isImportant = 'true'` to allow the
index page to focus on exclusively the best projects.